+++ app/main.js
@@ -341,6 +341,7 @@ async function initialize (isAppStart = true) {
     breakPlanner.clear()
     breakPlanner.appExclusionsManager.reinitialize(settings)
     breakPlanner.doNotDisturb(settings.get('monitorDnd'))
+    breakPlanner.fullscreenApp(settings.get('monitorFullscreenApp'))
     breakPlanner.naturalBreaks(settings.get('naturalBreaks'))
     breakPlanner.nextBreak()
   }
@@ -456,7 +457,8 @@ function onSuspendOrLock () {
   if (settings.get('pauseForSuspendOrLock')) {
     if (breakPlanner.isPaused || breakPlanner.dndManager.isOnDnd ||
       breakPlanner.naturalBreaksManager.isSchedulerCleared ||
-      breakPlanner.appExclusionsManager.isSchedulerCleared) {
+      breakPlanner.appExclusionsManager.isSchedulerCleared ||
+      breakPlanner.fullscreenAppManager.isOnFullscreenApp) {
       log.info('Stretchly: not pausing for suspendOrLock because paused already')
     } else {
       pausedForSuspendOrLock = true
@@ -511,7 +513,8 @@ function trayIconPath () {
       breakPlanner.isPaused ||
       breakPlanner.dndManager.isOnDnd ||
       breakPlanner.naturalBreaksManager.isSchedulerCleared ||
-      breakPlanner.appExclusionsManager.isSchedulerCleared,
+      breakPlanner.appExclusionsManager.isSchedulerCleared ||
+      breakPlanner.fullscreenAppManager.isOnFullscreenApp,
     monochrome: settings.get('useMonochromeTrayIcon'),
     inverted: settings.get('useMonochromeInvertedTrayIcon'),
     darkMode: nativeTheme.shouldUseDarkColors,
@@ -1212,6 +1215,8 @@ function pauseBreaks (milliseconds) {
 function resumeBreaks (notify = true) {
   if (breakPlanner.dndManager.isOnDnd) {
     log.info('Stretchly: not resuming breaks because in Do Not Disturb')
+  } else if (breakPlanner.fullscreenAppManager.isOnFullscreenApp) {
+    log.info('Stretchly: not resuming breaks because a fullscreen app is active')
   } else {
     breakPlanner.resume()
     log.info('Stretchly: resuming breaks')
@@ -1343,7 +1348,7 @@ function getTrayMenuTemplate () {
     return trayMenu
   }
 
-  if (!(breakPlanner.isPaused || breakPlanner.dndManager.isOnDnd || breakPlanner.appExclusionsManager.isSchedulerCleared)) {
+  if (!(breakPlanner.isPaused || breakPlanner.dndManager.isOnDnd || breakPlanner.appExclusionsManager.isSchedulerCleared || breakPlanner.fullscreenAppManager.isOnFullscreenApp)) {
     let submenu = []
     if (settings.get('microbreak')) {
       submenu = submenu.concat([{
@@ -1373,7 +1378,7 @@ function getTrayMenuTemplate () {
         updateTray()
       }
     })
-  } else if (!(breakPlanner.dndManager.isOnDnd || breakPlanner.appExclusionsManager.isSchedulerCleared)) {
+  } else if (!(breakPlanner.dndManager.isOnDnd || breakPlanner.appExclusionsManager.isSchedulerCleared || breakPlanner.fullscreenAppManager.isOnFullscreenApp)) {
     trayMenu.push({
       label: i18next.t('main.pause'),
       submenu: [
@@ -1515,6 +1520,10 @@ ipcMain.on('save-setting', function (event, key, value) {
 
   if (key === 'monitorDnd') {
     breakPlanner.doNotDisturb(value)
+  }
+
+  if (key === 'monitorFullscreenApp') {
+    breakPlanner.fullscreenApp(value)
   }
 
   if (key === 'language') {